# whatsapp-api/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"retryAfterSeconds": {
						"type": "number",
						"format": "double",
						"description": "Seconds until the client is expected to be ready (503 responses only)"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceUnavailableError": {
				"description": "503 Service Unavailable — WhatsApp client is not in the 'ready' state.\nShared across controllers so tsoa sees exactly one model definition.",
				"properties": {
					"error": {
						"type": "string"
					},
					"retryAfterSeconds": {
						"type": "number",
						"format": "double",
						"description": "Seconds until the client is expected to be ready (503 responses only)"
					}
				},
				"required": [
					"error",
					"retryAfterSeconds"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BadRequestError": {
				"description": "400 Bad Request — missing or invalid input fields.\nShared across controllers so tsoa sees exactly one model definition.",
				"properties": {
					"error": {
						"type": "string"
					},
					"retryAfterSeconds": {
						"type": "number",
						"format": "double",
						"description": "Seconds until the client is expected to be ready (503 responses only)"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ContactItem": {
				"properties": {
					"id": {
						"type": "string",
						"description": "WhatsApp contact ID"
					},
					"name": {
						"type": "string"
					},
					"pushname": {
						"type": "string"
					},
					"shortName": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"isGroup": {
						"type": "boolean"
					},
					"isWAContact": {
						"type": "boolean"
					},
					"isMyContact": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"pushname",
					"number",
					"isGroup",
					"isWAContact",
					"isMyContact"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ContactsResponse": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"contacts": {
						"items": {
							"$ref": "#/components/schemas/ContactItem"
						},
						"type": "array"
					}
				},
				"required": [
					"count",
					"contacts"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PhoneCheckResult": {
				"properties": {
					"whatsappId": {
						"type": "string",
						"description": "WhatsApp chat ID that was checked, e.g. \"15551234567@c.us\""
					},
					"registered": {
						"type": "boolean",
						"description": "Whether this number is registered on WhatsApp"
					}
				},
				"required": [
					"whatsappId",
					"registered"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PhoneCheckResponse": {
				"properties": {
					"input": {
						"type": "string",
						"description": "Original phone input as provided"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/PhoneCheckResult"
						},
						"type": "array",
						"description": "All candidate IDs that were checked (one if country code detected, two if not)"
					},
					"registered": {
						"type": "boolean",
						"description": "True if ANY candidate was found on WhatsApp"
					},
					"whatsappId": {
						"type": "string",
						"nullable": true,
						"description": "The first registered WhatsApp ID found, or null if none found"
					}
				},
				"required": [
					"input",
					"results",
					"registered",
					"whatsappId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveContactResponse": {
				"properties": {
					"ok": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"id": {
						"type": "string",
						"description": "WhatsApp ID of the saved contact (e.g. \"16073041892@c.us\")"
					},
					"phone": {
						"type": "string",
						"description": "The phone number that was saved"
					},
					"firstName": {
						"type": "string",
						"description": "Contact's first name"
					}
				},
				"required": [
					"ok",
					"id",
					"phone",
					"firstName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveContactBody": {
				"properties": {
					"phone": {
						"type": "string",
						"description": "Phone number in digits only (e.g. \"16073041892\"). Include country code."
					},
					"firstName": {
						"type": "string",
						"description": "Contact's first name"
					},
					"lastName": {
						"type": "string",
						"description": "Contact's last name (optional)"
					},
					"syncToAddressbook": {
						"type": "boolean",
						"description": "If true, also sync to the phone's address book. Default false."
					}
				},
				"required": [
					"phone",
					"firstName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-Api-Token",
				"description": "API_TOKEN from .env — set in X-Api-Token request header"
			},
			"basicAuth": {
				"type": "http",
				"scheme": "basic",
				"description": "BASIC_AUTH_USERNAME / BASIC_AUTH_PASSWORD from .env — standard HTTP Basic Auth"
			}
		}
	},
	"info": {
		"title": "WhatsApp Neo API",
		"version": "2.0.0",
		"description": "REST API for WhatsApp via whatsapp-web.js + remote Chromium CDP.\n\nAll routes except `/health` require either a Bearer token or Basic Auth:\n```\nX-Api-Token: <API_TOKEN>\nAuthorization: Basic <base64>\n```",
		"contact": {}
	},
	"paths": {
...
		"/contacts": {
			"get": {
				"operationId": "GetContacts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContactsResponse"
								}
							}
						}
					},
					"503": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceUnavailableError"
								}
							}
						}
					}
				},
				"description": "List all WhatsApp contacts saved on the connected account.",
				"tags": [
					"Contacts"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [
					{
						"description": "Maximum contacts to return (1–1000, default 200)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 200,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "SaveContact",
				"responses": {
					"201": {
						"description": "Contact saved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveContactResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"503": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceUnavailableError"
								}
							}
						}
					}
				},
				"description": "Save a new contact to the WhatsApp address book.\n\nThe `phone` field should be a number with country code (digits only, e.g.\n`\"16073041892\"`). The contact will be saved to the WhatsApp cloud and\noptionally synced to the phone's address book.",
				"tags": [
					"Contacts"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Contact details",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveContactBody",
								"description": "Contact details"
							}
						}
					}
				}
			}
		},
		"/contacts/check": {
			"get": {
				"operationId": "CheckPhone",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PhoneCheckResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"503": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceUnavailableError"
								}
							}
						}
					}
				},
				"description": "Check whether a phone number is registered on WhatsApp.\n\nIf no country code is detected (stripped input has ≤ 10 digits and no\nleading `+`), **both** `+1` (US/Canada) and `+52` (Mexico) are tried\nautomatically and all results are returned.\n\nExamples of accepted formats:\n- `+1 (555) 123-4567`  → country code detected, one check\n- `5551234567`          → 10 digits, no `+` → checks +1 and +52\n- `526441234567`        → 12 digits → country code detected, one check",
				"tags": [
					"Contacts"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [
					{
						"description": "Phone number in any format.",
						"in": "query",
						"name": "phone",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/contacts/{contactId}/avatar": {
			"get": {
				"operationId": "GetAvatar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string",
											"format": "byte"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"503": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceUnavailableError"
								}
							}
						}
					}
				},
				"description": "Get the profile picture (avatar) for a WhatsApp contact.\n\nProxies the profile picture from WhatsApp's CDN and returns the raw image\nbytes. Returns `404` if the contact has no profile picture set.\n\nThe frontend should call this endpoint lazily per visible row and fall back\nto showing coloured initials on `404` or network error.",
				"tags": [
					"Contacts"
				],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"basicAuth": []
					}
				],
				"parameters": [
					{
						"description": "WhatsApp contact ID, e.g. `16073041892@c.us`",
						"in": "path",
						"name": "contactId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		}
	],
...
}